requirements.txt
Cython==0.29.36
pandas==2.0.3
pmdarima==2.0.3
prophet==1.1.4
PyYAML==6.0
requests==2.31.0
urllib3==2.0.3